refactor: support for venv and more features

+++ src/isynclinux/sync_service.py
@@ -1,35 +1,51 @@
 import os
+import time
 from shutil import copyfileobj
 from pathlib import Path
 from shutil import copyfileobj
 from pyicloud import PyiCloudService
-from os.path import expanduser
+from os.path import expanduser, isfile
 from pyicloud.services.drive import DriveNode
 from os import stat
 
+from isynclinux.utils import config_dir
+
 
 class SyncService:
-    def __init__(self, api: PyiCloudService) -> None:
+    def __init__(self, api: PyiCloudService,file_list_cache_file, ignored_folders=[], verbose=False) -> None:
         self.api = api
         self.amount_of_files = 0
-        self.drive_size_mb = 0
+        self.drive_size_gb = 0
+        self.total_time_taken = 0
+        self.verbose = verbose
+        self.ignored_folders = ignored_folders
+        self.file_list_cache_file = file_list_cache_file
         return None
 
     def list_files_recursive(self, folder: DriveNode, current_path=""):
         icloud_file_list = []
+        start_time = time.time()
         files = folder.get_children()
+        end_time = time.time()
+        if self.verbose:
+            self.total_time_taken += end_time-start_time
+            print(f"Processing {current_path} ({end_time - start_time:.5f}s) ({self.total_time_taken:.5f}s)")
         for file in files:
             new_path = f"{current_path}/{file.name}" if current_path else file.name
+
             if file.type == "folder" or file.type == "app_library":
+                if file.name in self.ignored_folders:
+                    continue
                 files_in_subfolder = self.list_files_recursive(file, new_path)
                 icloud_file_list.extend(files_in_subfolder)
 
             elif file.type == "file":
                 icloud_file_list.append(new_path)
                 self.amount_of_files = self.amount_of_files + 1
-                file_size_mb = (file.size or 0) / 1000 / 1000
-                self.drive_size_mb = round(self.drive_size_mb + file_size_mb, 1)
-                print(f"\033[32mFound\033[0m {self.amount_of_files} files ({self.drive_size_mb} Mb)", end="\r")
+                file_size_gb = (file.size or 0) / 1000 / 1000 / 1000
+                if not self.verbose:
+                    print(f"\033[32mFound\033[0m {self.amount_of_files} files ({self.drive_size_gb:.4f} Gb)", end="\r")
+                self.drive_size_gb = round(self.drive_size_gb + file_size_gb, 4)
 
         return icloud_file_list
 
@@ -43,9 +59,16 @@ class SyncService:
         if root_folder is None:
             return 1
 
-        icloud_file_list = self.list_files_recursive(root_folder, "")
-        drive_size_mb = self.drive_size_mb
-        print(f"\033[32mFound\033[0m total of {len(icloud_file_list)} files ({drive_size_mb} Mb)")
+        icloud_file_list = []
+        if isfile(self.file_list_cache_file):
+            print("Found cached file list, do you want to use it?")
+            if input("y/N: ").lower() == "y":
+                icloud_file_list = self.read_cached_file_list()
+        if not icloud_file_list:
+            icloud_file_list = self.list_files_recursive(root_folder)
+            self.cache_file_list(icloud_file_list)
+            drive_size_gb = self.drive_size_gb
+            print(f"\033[32mFound\033[0m {len(icloud_file_list)} files ({drive_size_gb:.1f} Gb)")
 
         files_synced = 0
 
@@ -55,8 +78,9 @@ class SyncService:
             folders = file.split("/")
 
             percent_done = round(files_synced / len(icloud_file_list) * 100, 1)
-            pd = f"({percent_done} %)"
-            print(f"\033[32mSyncing\033[0m {files_synced} of {len(icloud_file_list)} {pd} {filler}", end="\r")
+            if not self.verbose:
+                pd = f"({percent_done} %)"
+                print(f"\033[32mSyncing\033[0m {files_synced} of {len(icloud_file_list)} {pd} {filler}", end="\r")
 
             if len(folders) == 1:
                 drive_file = root_folder[folders[0]]
@@ -67,6 +91,28 @@ class SyncService:
 
         print(f"\033[36mSyncing\033[0m completed! {filler}")
         return 0
+
+    def read_cached_file_list(self) -> list[str]:
+        try:
+            fp =open(self.file_list_cache_file, "r")
+            return [file.strip() for file in fp.readlines()]
+        except Exception as error:
+            print(error)
+            print(f"Couldn't open {self.file_list_cache_file}")
+            return []
+
+
+    def cache_file_list(self, files: list[str]):
+        try:
+            fp =open(self.file_list_cache_file, "w")
+            [fp.write(file + "\n") for file in files]
+            fp.close()
+
+        except:
+            print(f"Couldn't write: {self.file_list_cache_file}")
+            return
+
+
 
     def validate_dir_eligibility(self, sync_dir) -> int:
         if sync_dir == "":
@@ -106,10 +152,14 @@ class SyncService:
         except FileNotFoundError:
             pass  # file doesn't exist, download it
 
+        if self.verbose:
+            print(f"Syncing {drive_file}")
         self.write_local_file(drive_file, file_sync_dir)
 
     def write_local_file(self, drive_file, sync_dir):
-        response = drive_file.open(stream=True)
-        file_out = open(sync_dir + "/" + drive_file.name, 'wb')
-        copyfileobj(response.raw, file_out)
-        file_out.close()
+        local_file_path = sync_dir + "/" + drive_file.name
+        if not isfile(local_file_path):
+            response = drive_file.open(stream=True)
+            file_out = open(sync_dir + "/" + drive_file.name, 'wb')
+            copyfileobj(response.raw, file_out)
+            file_out.close()
